+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_sender_async.py
@@ -8,6 +8,7 @@ from typing import Any, TYPE_CHECKING, Union, List
 
 import uamqp
 from uamqp import SendClientAsync, types
+from uamqp.constants import MessageState
 
 from .._common.message import Message, BatchMessage
 from .._base_handler import _convert_connection_string_to_kwargs
@@ -278,6 +279,12 @@ class ServiceBusSender(BaseHandler, SenderMixin):
             pass
         if isinstance(message, BatchMessage) and len(message) == 0:  # pylint: disable=len-as-condition
             raise ValueError("A BatchMessage or list of Message must have at least one Message")
+        if not isinstance(message, BatchMessage):
+            if not isinstance(message, Message):
+                raise TypeError("Can only send azure.servicebus.<BatchMessage,Message> or lists of Messages.")
+            # This allows resending an already sent message; not needed for batch.  Perhaps should be in uamqp?
+            message.message.state = MessageState.WaitingToBeSent
+            message.message._response = None
 
         await self._do_retryable_operation(
             self._send,